Move utils into separate files
