+++ pybop/costs/error_measures.py
@@ -2,9 +2,9 @@ import numpy as np
 import pybamm
 
 from pybop._dataset import Dataset
-from pybop._utils import FailedSolution
 from pybop.costs.base_cost import BaseCost
 from pybop.parameters.parameter import Inputs
+from pybop.simulators.failed_solution import FailedSolution
 from pybop.simulators.solution import Solution
