+++ pybop/pybamm/__init__.py
@@ -2,3 +2,4 @@ from .simulator import Simulator
 from .eis_simulator import EISSimulator
 from .parameter_utils import set_formation_concentrations, cell_mass, cell_volume
 from .design_variables import add_variable_to_model
+from .utils import RecommendedSolver, SymbolReplacer
