+++ pybop/pybamm/simulator.py
@@ -8,9 +8,10 @@ from pybamm import SolverError
 if TYPE_CHECKING:
     from pybop.parameters.parameter import Inputs
 from pybop._dataset import Dataset
-from pybop._utils import FailedSolution, RecommendedSolver
 from pybop.parameters.parameter import Parameter, Parameters
+from pybop.pybamm.utils import RecommendedSolver
 from pybop.simulators.base_simulator import BaseSimulator
+from pybop.simulators.failed_solution import FailedSolution
 
 
 class Simulator(BaseSimulator):
@@ -38,7 +39,7 @@ class Simulator(BaseSimulator):
     protocol : pybamm.Experiment | Dataset | np.ndarray | None
         The protocol as an experiment, a 1D array of values or dataset containing (time) domain data.
     solver : pybamm.BaseSolver, optional
-        The solver to use to solve the model. If None, uses `pybop.RecommendedSolver`.
+        The solver to use to solve the model. If None, uses `pybop.pybamm.RecommendedSolver`.
     output_variables : list, optional
         A list of output variables to return.
     geometry : pybamm.Geometry, optional
